Implementação e testes de CompanyUser

+++ app/Http/Controllers/CompanyUserController.php
@@ -13,6 +13,8 @@ class CompanyUserController extends Controller
 {
     public function index(Company $company)
     {
+        $this->authorize('viewAny', $company);
+        
         $users = $company->users()->where('role_id', Role::COMPANY_OWNER->value)->get();
 
         return view('companies.users.index', compact('company', 'users'));
@@ -20,11 +22,15 @@ class CompanyUserController extends Controller
 
     public function create(Company $company)
     {
+        $this->authorize('create', $company);
+        
         return view('companies.users.create', compact('company'));
     }
 
     public function store(StoreUserRequest $request, Company $company)
     {
+        $this->authorize('create', $company);
+        
         $company->users()->create([
             'name' => $request->input('name'),
             'email' => $request->input('email'),
@@ -37,11 +43,15 @@ class CompanyUserController extends Controller
 
     public function edit(Company $company, User $user)
     {
+        $this->authorize('update', $company);
+        
         return view('companies.users.edit', compact('company', 'user'));
     }
 
     public function update(UpdateUserRequest $request, Company $company, User $user)
     {
+        $this->authorize('update', $company);
+        
         $user->update($request->validated());
 
         return to_route('companies.users.index', $company);
@@ -49,6 +59,8 @@ class CompanyUserController extends Controller
 
     public function destroy(Company $company, User $user)
     {
+        $this->authorize('delete', $company);
+        
         $user->delete();
 
         return to_route('companies.users.index', $company);
